refactor: todoId와 userId 타입 number로 변경,api 관련 함수 이름 및 위치 변경
